minibatches

+++ include/types.h
@@ -1,5 +1,5 @@
 #ifndef TYPES_H
 #define TYPES_H
-typedef float number;
+typedef double number;
 typedef unsigned long int index;
 #endif
